+++ src/models/torrent.rs
@@ -21,6 +21,7 @@
 use super::schema::*;
 use super::*;
 use super::{category::Category, user::User};
+use models::acl::Subject;
 
 #[derive(Debug, Queryable, Insertable, AsChangeset, Identifiable, Associations, Serialize)]
 #[table_name = "torrents"]
@@ -106,6 +107,10 @@ impl Torrent {
             .filter(t::id.eq(&self.id))
             .execute(db)
             .map_err(|e| format!("failed to delete torrent: {}", e).into())
+    }
+
+    pub fn comments(&self, db: &PgConnection) -> Vec<TorrentComment> {
+        TorrentComment::find_for_torrent(&self.id, db)
     }
 }
 
@@ -307,11 +312,7 @@ pub struct NewTorrentNFO<'a> {
 impl<'a> NewTorrentNFO<'a> {
     pub fn new(torrent_id: &'a Uuid, data: &'a [u8]) -> Self {
         let id = Uuid::new_v4();
-        NewTorrentNFO {
-            id,
-            torrent_id,
-            data,
-        }
+        NewTorrentNFO { id, torrent_id, data }
     }
 
     pub fn create(&self, db: &PgConnection) -> Result<usize> {
@@ -321,7 +322,7 @@ impl<'a> NewTorrentNFO<'a> {
     }
 }
 
-#[derive(Debug, Default)]
+#[derive(Default)]
 pub struct TorrentMsg {
     pub torrent: Torrent,
     pub torrent_user_name: Option<String>,
@@ -330,11 +331,12 @@ pub struct TorrentMsg {
     pub images: Vec<TorrentImage>,
     pub files: Vec<TorrentFile>,
     pub peers: Vec<(Peer, String)>,
+    pub comments: Vec<TorrentCommentResponse>,
     pub timezone: i32,
 }
 
 impl TorrentMsg {
-    pub fn find(id: &Uuid, db: &PgConnection) -> Result<Self> {
+    pub fn find(id: &Uuid, db: &PgConnection, subj: &UserSubject) -> Result<Self> {
         if let Some(torrent) = Torrent::find(id, db) {
             let nfo = TorrentNFO::find_for_torrent(id, db);
             let images = TorrentImage::find_for_torrent(id, db);
@@ -342,6 +344,7 @@ impl TorrentMsg {
             let peers = Peer::find_for_torrent(id, db);
             let torrent_user_name = torrent.user_name(db);
             let category = Category::find(&torrent.category_id, db).ok_or("category not found")?;
+            let comments = torrent.comments(db).into_iter().map(|c| TorrentCommentResponse::new(c, &db, &subj)).collect();
             let timezone = SETTINGS.read().unwrap().user.default_timezone;
 
             Ok(TorrentMsg {
@@ -352,6 +355,7 @@ impl TorrentMsg {
                 images,
                 files,
                 peers,
+                comments,
                 timezone,
             })
         } else {
@@ -379,6 +383,7 @@ pub struct TorrentList {
     pub last_action: Option<Timestamp>,
     pub last_seeder: Option<Timestamp>,
     pub created_at: Timestamp,
+    pub comments: i64,
 }
 
 impl TorrentList {
@@ -420,11 +425,7 @@ pub struct Transfer {
 }
 
 impl Transfer {
-    pub fn find_for_announce(
-        torrent_id: &Uuid,
-        user_id: &Uuid,
-        db: &PgConnection,
-    ) -> Option<Transfer> {
+    pub fn find_for_announce(torrent_id: &Uuid, user_id: &Uuid, db: &PgConnection) -> Option<Transfer> {
         transfers::table
             .filter(transfers::dsl::torrent_id.eq(torrent_id))
             .filter(transfers::dsl::user_id.eq(user_id))
@@ -432,16 +433,13 @@ impl Transfer {
             .ok()
     }
     pub fn save(&self, db: &PgConnection) -> Result<usize> {
-        let query = diesel::insert_into(transfers::table)
+        diesel::insert_into(transfers::table)
             .values(self)
             .on_conflict(on_constraint("transfers_user_id_torrent_id_key"))
             .do_update()
-            .set(self);
-        trace!(
-            "query: {}",
-            diesel::debug_query::<diesel::pg::Pg, _>(&query)
-        );
-        query.execute(db).chain_err(|| "transfer update failed")
+            .set(self)
+            .execute(db)
+            .chain_err(|| "transfer update failed")
     }
 }
 
@@ -509,4 +507,131 @@ impl<'a> NewTorrentImage<'a> {
             .execute(db)
             .map_err(|e| format!("torrent images insert failed: {}", e).into())
     }
+}
+
+#[derive(Queryable, Identifiable, Associations, Insertable, Serialize)]
+#[belongs_to(User)]
+#[belongs_to(Torrent)]
+pub struct TorrentComment {
+    id: Uuid,
+    user_id: Uuid,
+    torrent_id: Uuid,
+    content: String,
+    created_at: Timestamp,
+    updated_at: Timestamp,
+}
+
+impl TorrentComment {
+    pub fn find(id: &Uuid, db: &PgConnection) -> Option<Self> {
+        use schema::torrent_comments::dsl;
+        dsl::torrent_comments.find(id).first::<Self>(db).ok()
+    }
+
+    pub fn find_for_torrent(torrent_id: &Uuid, db: &PgConnection) -> Vec<Self> {
+        use schema::torrent_comments::dsl;
+        dsl::torrent_comments
+            .filter(dsl::torrent_id.eq(torrent_id))
+            .order_by(dsl::created_at.asc())
+            .load::<Self>(db)
+            .unwrap_or_default()
+    }
+
+    pub fn set_content(&mut self, content: String) -> String {
+        std::mem::replace(&mut self.content, content)
+    }
+
+    pub fn save(&mut self, db: &PgConnection) -> Result<usize> {
+        use schema::torrent_comments::dsl as t;
+        self.updated_at = Utc::now();
+        diesel::update(schema::torrent_comments::table)
+            .set(self.update_set())
+            .filter(t::id.eq(&self.id))
+            .execute(db)
+            .map_err(|e| format!("failed to update torrent comment: {}", e).into())
+    }
+
+    fn update_set<'a>(&'a self) -> UpdateTorrentComment<'a> {
+        UpdateTorrentComment {
+            content: &self.content[..],
+            updated_at: &self.updated_at,
+        }
+    }
+
+    pub fn delete(&self, db: &PgConnection) -> Result<usize> {
+        use schema::torrent_comments::dsl as t;
+        diesel::delete(schema::torrent_comments::table)
+            .filter(t::id.eq(&self.id))
+            .execute(db)
+            .map_err(|e| format!("failed to delete comment: {}", e).into())
+    }
+}
+
+impl HasUser for TorrentComment {
+    fn user_id(&self) -> &Uuid {
+        &self.user_id
+    }
+}
+
+#[derive(Serialize)]
+pub struct TorrentCommentResponse {
+    #[serde(flatten)]
+    comment: TorrentComment,
+    user_name: String,
+    may_edit: bool,
+    may_delete: bool,
+}
+
+impl TorrentCommentResponse {
+    pub fn new(comment: TorrentComment, db: &PgConnection, subj: &UserSubject) -> Self {
+        let user_name = comment.user_name(db);
+        let may_edit = subj.may_write(&comment);
+        let may_delete = subj.may_delete(&comment);
+        Self { comment, user_name, may_edit, may_delete }
+    }
+
+    pub fn new2(comment: TorrentComment, db: &PgConnection) -> Self {
+        let user_name = comment.user_name(db);
+        let may_edit = false;
+        let may_delete = false;
+        Self { comment, user_name, may_edit, may_delete }
+    }
+
+    pub fn comment(&self) -> &TorrentComment {
+        &self.comment
+    }
+}
+
+#[derive(Insertable, Identifiable)]
+#[table_name = "torrent_comments"]
+pub struct NewTorrentComment<'a> {
+    id: Uuid,
+    user_id: &'a Uuid,
+    torrent_id: &'a Uuid,
+    content: &'a str,
+}
+
+impl<'a> NewTorrentComment<'a> {
+    pub fn new(user_id: &'a Uuid, torrent_id: &'a Uuid, content: &'a str) -> Self {
+        let id = Uuid::new_v4();
+        Self {
+            id,
+            user_id,
+            torrent_id,
+            content,
+        }
+    }
+
+    pub fn create(&self, db: &PgConnection) -> Result<TorrentComment> {
+        diesel::insert_into(schema::torrent_comments::table)
+            .values(self)
+            .get_result::<TorrentComment>(db)
+            .map_err(|e| format!("torrent comment insert failed: {}", e).into())
+    }
+}
+
+#[derive(AsChangeset)]
+#[table_name = "torrent_comments"]
+pub struct UpdateTorrentComment<'a> {
+    content: &'a str,
+    updated_at: &'a Timestamp,
 }